Calculate the study and arm level survivor functions estimates from a 2nd order fractional polynomial NMA. These estimates provide the basis for a goodness-of-fit graph when plotted along with the input data.

Calculate the study and arm level survivor functions estimates from a 2nd order fractional polynomial NMA. These estimates provide the basis for a goodness-of-fit graph when plotted along with the input data.

get_fp_2o_GoF(fit, time = 1:24, bl.node = "mu", contrast.node = "d")

Arguments

fit
JAGS object with FP NMA fit.

time
Vector of time-points at which S(t) functions are calculated.

bl.node
Charactor to identify the node in the jags model that identifies the baseline estimates (default is "mu").

contrast.node
Charactor to identify the node in the jags model that identifies the baseline estimates (default is "d").

Value

a data.frame containing survivor function for each treatment

Details

The FP parameters to calculate the log-hazard curves for each survivor function are combinations bl.node + contrast.node. Here, every study-specific baseline estimate is combined with the arm-specific contrast to obtain arm level NMA estimates.

Examples