Utility function for pre-processing: prepare jags input data for PWE model.

Utility function for pre-processing: prepare jags input data for PWE model.

groupedTTE_pwe_pre_proc(
  dat,
  ref.std = NULL,
  nma.ref.trt = NULL,
  model.pars = NULL,
  feprior_mean = NULL,
  feprior_prec = NULL,
  bth.prior = NULL
)

Arguments

dat

A data.frame with columns study, treatment, t.start, t.end, n.event, n.risk.

ref.std

A character of the reference study

nma.ref.trt

A character of the reference treatment

model.pars

list, containing exponents (numeric vector of exponents for FP model) and t.eval - one of "midpoint" (the default), "start", or "end", to identify the time-point within each interval used to calculate the hazard rate (over each interval, the hazard function is approximated with this value h(t.eval)); alternatively, a numeric value in the interval [0,1] used to define t.eval (where for example "midpoint" corresponds to the value 0.5).

feprior_mean

A numeric value representing feprior_mean (default 0)

feprior_prec

A numeric value representing feprior_prec default 0.0001

bth.prior

A numeric list containing between-trial heterogeneity priors

Value

A list with input data for jags fit of PWE model.

See also