Creates a fractional polynomial model plan

Creates a fractional polynomial model plan

plan_pwe(
  model.pars,
  bth.model = c("FE", "RE"),
  ref.std,
  nma.ref.trt,
  engine = "rjags",
  data.type = "GSD",
  feprior_mean = 0,
  feprior_prec = 1e-04,
  descr = "Piecewise Exponential model",
  descr_s = "PWE",
  n.chains = NULL,
  n.iter = NULL,
  n.burnin = NULL,
  n.thin = NULL,
  bth.prior = NULL,
  rsd = 13579,
  model.file = NULL,
  ...
)

Arguments

model.pars

A named list containing `cut.pts` when using GSD

bth.model

One of "FE" or "RE"

ref.std

A string containing the reference study

nma.ref.trt

A string containing the reference treatment

engine

Only `rjags` currently supported

data.type

Only "GSD" supported

feprior_mean

A numeric value representing feprior_mean (default 0)

feprior_prec

A numeric value representing feprior_prec default 0.0001

descr

A character string to describe the model

descr_s

Short description

n.chains

A numeric value representing n.chains

n.iter

A numeric value representing n.iter

n.burnin

A numeric value representing n.burnin

n.thin

A numeric value representing n.thin

bth.prior

A list containing type & distr. Only required if bth.model == "FE"

rsd

A numeric representing seed to be set

model.file

Path to BUGS file. If NULL file name will be built from input parameters and extracted from BUGScode folder if present

...

any other named arguments must match arguments names from dic.samples or jags functions

Value

A list containing model, engine, analysis and model params

Examples

 
model_plan <- plan_pwe(model.pars = list(cut.pts =  c(3, 10)),
                       bth.model = "FE", ref.std = "STUDY2", nma.ref.trt = "B",                        
                       n.chains = 2,
                       n.iter = 6000,
                       n.burnin = 1000,
                       n.thin = 1)